+++ src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod base;
+pub mod dotenv;
 pub mod http;
 pub mod io;
 pub mod json;