+++ src/api/mod.rs
@@ -5,3 +5,4 @@ pub mod io;
 pub mod json;
 pub mod os;
 pub mod toml;
+pub mod yaml;