+++ src/main.rs
@@ -8,7 +8,7 @@ mod utils;
 #[cfg(windows)]
 mod windows_utf8;
 
-pub const VERSION: &str = "0.1.11";
+pub const VERSION: &str = "0.1.12";
 
 use crate::helper::print::{
     END,