+++ src/script.rs
@@ -2,7 +2,7 @@ use mlua::{Lua, Result, Table};
 use std::fs;
 use std::path::Path;
 
-use crate::api::{base, io as api_io, os as api_os, http as api_http, json as api_json, toml as api_toml, dotenv as api_dotenv};
+use crate::api::{base, io as api_io, os as api_os, http as api_http, json as api_json, toml as api_toml, dotenv as api_dotenv, yaml as api_yaml};
 
 pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
     if !Path::new(file).exists() {
@@ -27,6 +27,7 @@ pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
     let dapi_json = api_json::register(&lua)?;
     let dapi_toml = api_toml::register(&lua)?;
     let dapi_dotenv = api_dotenv::register(&lua)?;
+    let dapi_yaml = api_yaml::register(&lua)?;
 
     let globals = lua.globals();
     let package: Table = globals.get("package")?;
@@ -39,6 +40,7 @@ pub fn execute_script(file: &str, safe_mode: &bool) -> Result<()> {
     preload.set("dapi_json", lua.create_function(move |_, ()| Ok(dapi_json.clone()))?)?;
     preload.set("dapi_toml", lua.create_function(move |_, ()| Ok(dapi_toml.clone()))?)?;
     preload.set("dapi_dotenv", lua.create_function(move |_, ()| Ok(dapi_dotenv.clone()))?)?;
+    preload.set("dapi_yaml", lua.create_function(move |_, ()| Ok(dapi_yaml.clone()))?)?;
 
     lua.load(&script).exec()?;
     Ok(())