+++ src/utils/mod.rs
@@ -3,5 +3,4 @@ pub mod ini_utils;
 pub mod zip_utils;
 pub mod toml_utils;
 pub mod xml_utils;
-pub mod yaml_utils;
 pub mod json_utils;